requirements.txt
amadeus==12.0.0
asgiref==3.8.1
astroid==3.3.11
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
dill==0.4.0
dj-database-url==3.0.1
Django==5.2.3
django-environ==0.12.0
djangorestframework==3.16.1
git-filter-repo==2.47.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google_search_results==2.4.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
numpy==2.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
platformdirs==4.3.8
playwright==1.52.0
playwright-stealth==1.0.6
proto-plus==1.26.1
protobuf==6.32.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyee==13.0.0
pylint==3.3.8
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
rsa==4.9.1
selenium==4.33.0
six==1.17.0
snakeviz==2.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.43
sqlparse==0.5.3
tomlkit==0.13.3
tornado==6.4
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
